// src-ui/src/lib/state_machine/generated/StatechartDeltaDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { StateUiNodeDto } from "./StateUiNodeDto";

export type StatechartDeltaDto = { "kind": "upsert", state: StateUiNodeDto, } | { "kind": "remove", state_id: string, } | { "kind": "active_changed", state_id: string, active: boolean, };