// src-ui/src/lib/state_machine/generated/StateUiNodeDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { StateUiKind } from "./StateUiKind";
import type { StateUiLayoutDto } from "./StateUiLayoutDto";

export type StateUiNodeDto = { id: string, label: string, active: boolean, kind: StateUiKind, layout: StateUiLayoutDto, child_region_ids: Array<string>, };